Add param for disabling pod cleanup (#2903)
Introduces a flow parameter `disable.pod.cleanup`, if it's set
the flow-pod will not be auto-cleaned after flow execution ends.

+++ azkaban-common/src/main/java/azkaban/executor/container/watch/FlowStatusManagerListener.java
@@ -266,6 +266,11 @@ public class FlowStatusManagerListener implements AzPodStatusListener {
    */
   private void deleteFlowContainer(AzPodStatusMetadata event) {
     logger.info("Deleting Flow Pod: " + event.getPodName());
+    if (event.getFlowPodMetadata().get().isCleanupDisabled()) {
+      logger.warn(format("Pod deletion is disabled for pod %s, please delete it manually.",
+          event.getPodName()));
+      return;
+    }
     try {
       containerizedImpl.deleteContainer(
           Integer.parseInt(
